+++ amplipi/defaults.py
@@ -10,6 +10,7 @@ from amplipi import models
 
 MUTE_ALL_ID = 10000
 LAST_PRESET_ID = 9999
+AUX_STREAM_ID = 995
 RCAs = [996, 997, 998, 999]
 LMS_DEFAULTS = [1000, 1001, 1002, 1003]
 
@@ -25,6 +26,7 @@ DEFAULT_CONFIG = {  # This is the system state response that will come back from
   # NOTE: streams and groups seem like they should be stored as dictionaries with integer keys
   #       this does not make sense because JSON only allows string based keys
   "streams": [
+    {"id": AUX_STREAM_ID, "name": "Aux", "type": "aux", "disabled": False},
     {"id": RCAs[0], "name": "Input 1", "type": "rca", "index": 0, "disabled": False},
     {"id": RCAs[1], "name": "Input 2", "type": "rca", "index": 1, "disabled": False},
     {"id": RCAs[2], "name": "Input 3", "type": "rca", "index": 2, "disabled": False},
@@ -73,6 +75,7 @@ STREAMER_CONFIG = {  # This is the system state response that will come back fro
     {"id": 3, "name": "Output 4", "input": ""},
   ],
   "streams": [
+    {"id": AUX_STREAM_ID, "name": "Aux", "type": "aux", "disabled": False},
     {"id": 1000, "name": "Groove Salad", "type": "internetradio", "url": "http://ice6.somafm.com/groovesalad-32-aac",
      "logo": "https://somafm.com/img3/groovesalad-400.jpg", "disabled": False},
   ],
@@ -94,6 +97,7 @@ DEFAULT_LMS_CONFIG = {  # This is the system state response that will come back 
   # NOTE: streams and groups seem like they should be stored as dictionaries with integer keys
   #       this does not make sense because JSON only allows string based keys
   "streams": [
+    {"id": AUX_STREAM_ID, "name": "Aux", "type": "aux", "disabled": False},
     {"id": RCAs[0], "name": "Input 1", "type": "rca", "index": 0, "disabled": False},
     {"id": RCAs[1], "name": "Input 2", "type": "rca", "index": 1, "disabled": False},
     {"id": RCAs[2], "name": "Input 3", "type": "rca", "index": 2, "disabled": False},
@@ -143,6 +147,7 @@ STREAMER_LMS_CONFIG = {  # This is the system state response that will come back
     {"id": 4, "name": "Output 4", "input": f"stream={LMS_DEFAULTS[3]}"},
   ],
   "streams": [
+    {"id": AUX_STREAM_ID, "name": "Aux", "type": "aux", "disabled": False},
     {"id": LMS_DEFAULTS[0], "name": "Music 1", "type": "lms", "server": "localhost"},
     {"id": LMS_DEFAULTS[1], "name": "Music 2", "type": "lms", "server": "localhost"},
     {"id": LMS_DEFAULTS[2], "name": "Music 3", "type": "lms", "server": "localhost"},